style: reformat with 2-space indent
I'd only changed it to 4-space in order to be consistent with what I use
every day at work. Now we can go back to the comfort of The Old Ways
(tm).

+++ fig/posix/rm.ts
@@ -5,34 +5,33 @@ import run from '../run.js';
 import stringify from '../stringify.js';
 
 type Options = {
-    recurse?: boolean;
-    sudo?: boolean;
+  recurse?: boolean;
+  sudo?: boolean;
 };
 
 export default async function rm(
-    path: string,
-    options: Options = {}
+  path: string,
+  options: Options = {}
 ): Promise<Error | null> {
-    const passphrase = options.sudo ? await Context.sudoPassphrase : undefined;
+  const passphrase = options.sudo ? await Context.sudoPassphrase : undefined;
 
-    const args = ['-f', path];
+  const args = ['-f', path];
 
-    if (options.recurse) {
-        args.unshift('-r');
-    }
+  if (options.recurse) {
+    args.unshift('-r');
+  }
 
-    log.debug(`Removing: ${args.join(' ')}`);
+  log.debug(`Removing: ${args.join(' ')}`);
 
-    const result = await run('rm', args, {passphrase});
+  const result = await run('rm', args, {passphrase});
 
-    if (result.status === 0) {
-        return null;
-    } else {
-        log.debug(stringify(result));
+  if (result.status === 0) {
+    return null;
+  } else {
+    log.debug(stringify(result));
 
-        return (
-            result.error ||
-            new ErrorWithMetadata(`\`rm ${args.join(' ')}\` failed`)
-        );
-    }
+    return (
+      result.error || new ErrorWithMetadata(`\`rm ${args.join(' ')}\` failed`)
+    );
+  }
 }
